requirements.txt
streamlit
numpy
pandas
joblib
scikit-learn
langchain
langgraph
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-groq
python-dotenv
faiss-cpu
sentence-transformers
huggingface_hub
